requirements.txt
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.2
graphviz==0.13.2
-e git+https://github.com/openai/gym@21bd4f304be8f5cf7d0354e50ae4a977b7cfe5eb#egg=gym
gym-pull==0.1.7
-e git+https://github.com/openai/retro.git@46c70775cce5334de241b4094eac65ef9f0dd3d4#egg=gym_retro
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
neat-python==0.92
numpy==1.17.3
opencv-python==4.1.1.26
pyglet==1.3.2
pyparsing==2.4.4
python-dateutil==2.8.1
requests==2.22.0
scipy==1.3.1
six==1.13.0
urllib3==1.25.6
